refactor: contact package
