+++ backend/src/main/java/org/tarjetamish/contact/repository/JdbcContactRepository.java
@@ -5,7 +5,7 @@ import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.stereotype.Repository;
 import org.tarjetamish.contact.model.Contact;
 import org.tarjetamish.contact.repository.impl.ContactRepository;
-import org.tarjetamish.contact.mapper.ContactRowMapper;
+import org.tarjetamish.contact.mapper.impl.ContactRowMapper;
 
 import java.util.List;
 import java.util.Optional;
@@ -15,28 +15,29 @@ import java.util.Optional;
 public class JdbcContactRepository implements ContactRepository {
 
     private final JdbcTemplate jdbc;
+    private final ContactRowMapper contactRowMapper;
     @Override
     public List<Contact> findAll() {
         String sql = "SELECT * FROM tarjeta_mish.contact";
-        return jdbc.query(sql, new ContactRowMapper());
+        return jdbc.query(sql, contactRowMapper);
     }
 
     @Override
     public Optional<Contact> findByName(String name) {
         String sql = "SELECT * FROM tarjeta_mish.contact WHERE name = ?";
-        return Optional.empty();
+        return Optional.ofNullable(jdbc.queryForObject(sql, contactRowMapper, name));
     }
 
     @Override
     public Optional<Contact> findByAlias(String alias) {
         String sql = "SELECT * FROM tarjeta_mish.contact WHERE alias = ?";
-        return Optional.empty();
+        return Optional.ofNullable(jdbc.queryForObject(sql, contactRowMapper, alias));
     }
 
     @Override
     public Contact save(Contact contact) {
         String sql = "INSERT INTO tarjeta_mish.contact (name, numbaccount, email, alias, typeaccount, bank, iduser) VALUES (?, ?, ?, ?, ?, ?, ?)";
-        return jdbc.queryForObject(sql, new ContactRowMapper(), contact.getName(), contact.getAccountNumber(), contact.getEmail(), contact.getAlias(), contact.getTypeAccount().name(), contact.getBank().name(), contact.getIdUser());
+        return jdbc.queryForObject(sql, contactRowMapper, contact.getName(), contact.getAccountNumber(), contact.getEmail(), contact.getAlias(), contact.getTypeAccount().name(), contact.getBank().name(), contact.getIdUser());
     }
 
     @Override
